+++ src/com/kaylerrenslow/armaplugin/ArmaAddonsIndexingCallback.java
@@ -3,9 +3,14 @@ package com.kaylerrenslow.armaplugin;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.io.File;
+
 /**
- * Contains a set of methods to track the progress of Arma Addon/Mod indexing progress from {@link ArmaAddonsManager#loadAddons(ArmaAddonsProjectConfig)}.
- * Each method, besides {@link #finishedXMLParse()} and {@link #cancelAll()}, are invoked for one particular addon.
+ * Contains a set of methods to track the progress of Arma Addon/Mod indexing progress from {@link ArmaAddonsManager#loadAddonsAsync(ArmaAddonsProjectConfig, File, ArmaAddonsIndexingCallback)}.
+ * Each method is invoked for one particular addon.
+ * <p>
+ * All methods will <b>not</b> run on the JavaFX Thread of Java Swing Thread.
+ * Instead, the methods are invoked from a temporary thread.
  *
  * @author Kayler
  * @since 10/26/2017
@@ -36,23 +41,12 @@ public interface ArmaAddonsIndexingCallback {
 	}
 
 	/**
-	 * This is invoked when the addons config xml file has been parsed and interpreted.
-	 * This method will be invoked only once.
-	 */
-	void finishedXMLParse();
-
-	/**
 	 * This is invoked when a mod is beginning to get indexed. A {@link ArmaAddonIndexingHandle} is passed that contains
 	 * methods to cancel the indexing of a specific addon.
 	 *
 	 * @param handle a handle for the addon
 	 */
 	void workBegin(@NotNull ArmaAddonIndexingHandle handle);
-
-	/**
-	 * @return true to cancel all indexing for all addons, or false to keep them running
-	 */
-	boolean cancelAll();
 
 	/**
 	 * Invoked when the total overall work for an addon has been updated.
@@ -103,12 +97,21 @@ public interface ArmaAddonsIndexingCallback {
 	void warningMessage(@NotNull ArmaAddonIndexingHandle handle, @NotNull String message, @Nullable Exception e);
 
 	/**
-	 * Invoked when the indexing of an addon has reached a new step
+	 * Invoked when the indexing of an addon has reached and started a new step (this step may not ever be finished)
 	 *
 	 * @param handle  the addon's handle
 	 * @param newStep the new step
 	 */
-	void stepChange(@NotNull ArmaAddonIndexingHandle handle, @NotNull Step newStep);
+	void stepStart(@NotNull ArmaAddonIndexingHandle handle, @NotNull Step newStep);
+
+	/**
+	 * Invoked when the indexing of an addon has finished a step
+	 *
+	 * @param handle       the addon's handle
+	 * @param stepFinished the new step
+	 */
+	void stepFinish(@NotNull ArmaAddonIndexingHandle handle, @NotNull Step stepFinished);
+
 
 	/**
 	 * Invoked when the addon is completed indexed and the total work