+++ src/com/kaylerrenslow/armaplugin/ArmaPluginProjectConfigurable.java
@@ -4,11 +4,9 @@ import com.intellij.openapi.options.Configurable;
 import com.intellij.openapi.options.ConfigurationException;
 import com.intellij.openapi.util.io.FileUtil;
 import com.kaylerrenslow.armaplugin.dialog.ArmaPluginSettingsForm;
-import com.kaylerrenslow.armaplugin.dialog.IndexArmaAddonsStatusDialog;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.Nullable;
 
-import javax.swing.JButton;
 import javax.swing.JComponent;
 import java.io.File;
 
@@ -33,11 +31,11 @@ public class ArmaPluginProjectConfigurable implements Configurable {
 		form.initArma3ToolsDirectory(a3ToolsDir == null ? "" : a3ToolsDir.getAbsolutePath());
 
 		//testing code
-		JButton btnTest = new JButton("Test");
-		form.getPanelRoot().add(btnTest);
-		btnTest.addActionListener(e -> {
-			new IndexArmaAddonsStatusDialog().setVisible(true);
-		});
+//		JButton btnTest = new JButton("Test");
+//		form.getPanelRoot().add(btnTest);
+//		btnTest.addActionListener(e -> {
+//			new IndexArmaAddonsStatusDialog().setVisible(true);
+//		});
 
 		//end testing code
 		return form.getPanelRoot();