+++ CHANGES.md
@@ -1,11 +1,12 @@
 **Added**  
-* start for setting arma tools directory
+* 
 
 **Changed**  
-* moved some files around into more suitable packages
+* made ArmaAddonsManager utilize ArmaAddonsIndexingCallback.java and ArmaAddonIndexingHandle
 
 **Fixed**  
 * 
 
 **Notes**
 * scope works, but it doesn't work for deep nested scopes
+* ArmaAddonsManager needs to finish ArmaAddonsIndexingCallback.java utilization by incrementing an addon's current work progress and total work progress